% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opl_package.R
\name{OPL}
\alias{OPL}
\title{OPL: Optimal Policy Learning Package}
\description{
The OPL package provides tools for estimating and optimizing policy assignment rules
based on machine learning and econometric techniques.
}
\section{Main functions}{

\itemize{
\item \code{make_cate()}: Computes conditional average treatment effects.
\item \code{opl_tb()}: Optimal policy learning for threshold-based policies.
\item \code{opl_lc()}: Optimal policy learning for linear combination policies.
\item \code{opl_dt()}: Optimal policy learning for decision tree-based policies.
}
}

\section{Installation}{

To install the package from CRAN, use:
\code{install.packages("OPL")}
}

\section{Acknowledgments}{

The development of this software was supported by FOSSR (Fostering Open Science in Social Science Research), a project funded by the European Union - NextGenerationEU under the NPRR Grant agreement n. MURIR0000008.
}

\keyword{"_PACKAGE"}
\keyword{internal}
