% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opl_dt_max_choice.R
\name{opl_dt_max_choice}
\alias{opl_dt_max_choice}
\title{User selection on multiple choice}
\usage{
opl_dt_max_choice(nc, col_max, verbose = TRUE)
}
\arguments{
\item{nc}{Numeber of max welfare.}

\item{col_max}{Row index for max constrained welfare.}

\item{verbose}{Set TRUE to print the output on the console.}
}
\value{
Return the user's selection as an input.
}
\description{
Function that allows the user to select a row of maximum welfare among the rows with maximum welfare constrained. The function prints out the result and requires user input to select the row.
}
