% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{overlapping}
\alias{overlapping}
\title{Testing overlap between old and new policy sample}
\usage{
overlapping(train_data, test_data, x)
}
\arguments{
\item{train_data}{Train Dataset indicating the old policy sample.}

\item{test_data}{Test Dataset indicating the new policy sample.}

\item{x}{Vector of predictor variables.}
}
\value{
The function prints the superposition graph and the results of the Kolmogorov-Smirnov test.
}
\description{
Function to perform overlap analysis between train and test datasets.
The function performs principal component analysis (PCA) on the covariates for both sets
and calculates the Kolmogorov-Smirnov test for overlap.
}
